refactor: Rename the concrete implementation of module from `module` to `directoryModule`
